53강부터 들을 차례

+++ src/main/java/com/example/fileupload/article_image/dao/ArticleImageRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ArticleImageRepository extends JpaRepository<ArticleImage, Long> {
-    List<ArticleImage> findByIdIn(List<Integer> imageIdList);
+    List<ArticleImage> findByIdIn(List<Long> imageIdList);
 
     List<ArticleImage> findByArticle(Article article);
 }
